feat: move to microbundle-based build
Basically microbundle does everything we need to build lib like reflect
