+++ .gitignore
@@ -4,6 +4,7 @@ node_modules
 yarn-error.log
 
 /dist
+/ssr/dist
 /typings
 /browser
 /no-ssr
