+++ dist-test/no-ssr/reflect.test.tsx
@@ -5,7 +5,7 @@ import { act } from 'react-dom/test-utils';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { reflect } from '../../dist/esm';
+import { reflect } from '../../index';
 
 // Example1 (InputCustom)
 const InputCustom: FC<{
