+++ dist-test/ssr/create-reflect.test.tsx
@@ -4,7 +4,7 @@ import { Provider } from 'effector-react/ssr';
 
 import { render } from '@testing-library/react';
 
-import { createReflect } from '../../dist/esm/ssr';
+import { createReflect } from '../../ssr';
 
 // Example1 (InputCustom)
 const InputCustom: FC<{
